requirements.txt
Office365-REST-Python-Client==2.3.16
openai==1.3.6
streamlit==1.31.1
python-dotenv==1.0.1
nest-asyncio==1.6.0
llama_index==0.9.39
pymongo==4.6.1
pymupdf==1.23.19
weaviate-client==4.4.4
streamlit-option-menu==0.3.12